server: notifications will also go to the client that triggered it
This applies to the following notifications that go to all clients on
the allow-list for a room:

* `room created`
* `room active`
* `room destroyed`

This reverts previous attempts at filtering the "trigger client" from
this notification. It simplifies the logic in the client. Also update
the protocol doc accordingly.

Notifications that only go to room members behave the same as before,
because they must.