Allow accepting both TLS and insecure connections
We are moving to using nginx as a proxy for the server, to fix some
IPv6 SNI issues with ISPs that use a NAT64. Nginx will handle TLS for
us, and should fix that.

However, we have a tagged alpha release that we shouldn't break, so
now we need to accept both insecure (on localhost) and TLS connections
(on a public interface).